aBasic test for NewsList
