+++ server.js
@@ -62,7 +62,7 @@ app.post("/login", (req, res, next) => {
   }
   const payload = { id: "777", username };
   const token = jwt.sign(payload, secretJWTKeyThatShouldNotBeHere, {
-    expiresIn: "24h" // should expire in 1 hour on prod ###TODO 60 * 60
+    expiresIn: 60 * 60
   });
   res.status(200).send(token);
 });
