+++ src/redux/store.js
@@ -1,0 +1,18 @@
+import { createStore, applyMiddleware } from "redux";
+import { composeWithDevTools } from "redux-devtools-extension";
+import thunk from "redux-thunk";
+
+
+import rootReducer from "./rootReducer";
+
+const initialState = { 
+  user: localStorage.getItem("JWT") || null
+};
+
+const store = createStore(
+  rootReducer,
+  initialState,
+  composeWithDevTools(applyMiddleware(thunk))
+);
+
+export default store;
